requirements.txt
# Core Web Framework
flask==3.0.0
python-dotenv==1.0.0

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# AI APIs
anthropic==0.8.1
openai==1.6.1

# Data Processing
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Database
sqlite3  # Built into Python
sqlalchemy==2.0.23

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Optional: Production Dependencies
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.9